263 - Improve how videojs is loaded (#327)
